Restructure project
- Rename Models -> Entities
- Move ComponentShopDBContext to Repositories folder
